gsfdgdfgsdfgsdf

+++ index.js
@@ -97,6 +97,7 @@ client.on('message', message => {
             .setTitle('**POKUTA**')                                         // Nadpis
             .setURL('https://www.youtube.com/watch?v=dQw4w9WgXcQ')                          //Co se otevře po kliknutí na nadpis 
             .setAuthor('Policie České republiky', 'https://media.discordapp.net/attachments/729654155798380548/769906915022471208/pcr.png')    //Horní část embedu - Kdo to vytvoři, obrázek vpravu, co se otevře po kliknutí
+            .setDescription(`Pokutu vystavil ${message.member.nickname} a zapsal jí do databáze PČR`)
             .setThumbnail('https://upload.wikimedia.org/wikipedia/commons/thumb/8/8d/Paper-notes.svg/512px-Paper-notes.svg.png') //Obrázek vlevo
             .addFields(
                 { name: '**OSOBA**', value: jmeno_promene[1], inline: false },
@@ -104,7 +105,25 @@ client.on('message', message => {
                 { name: '**DŮVOD**', value: jmeno_promene[3], inline: false },
             )          
             .setTimestamp()
-            .setFooter(`Pokutu vystavil ${message.member.nickname} a zapsal jí do databáze PČR`, 'https://media.discordapp.net/attachments/729654155798380548/769906915022471208/pcr.png');           // URL obrázku dole
+            //.setFooter(`Pokutu vystavil ${message.member.nickname} a zapsal jí do databáze PČR`, 'https://media.discordapp.net/attachments/729654155798380548/769906915022471208/pcr.png');           // URL obrázku dole
+        message.channel.send(exampleEmbed)
+        message.delete()
+    }
+    if(jmeno_promene[0] === '!vazbaxxx'){
+        const exampleEmbed = new Discord.MessageEmbed()
+            .setColor('#FF0000')
+            .setTitle('**ODNĚTÍ SVOBODY*')                                         // Nadpis
+            .setURL('https://www.youtube.com/watch?v=dQw4w9WgXcQ')                          //Co se otevře po kliknutí na nadpis 
+            .setAuthor('Policie České republiky', 'https://media.discordapp.net/attachments/729654155798380548/769906915022471208/pcr.png')    //Horní část embedu - Kdo to vytvoři, obrázek vpravu, co se otevře po kliknutí
+            .setDescription(`Policista ${message.member.nickname} udělil odnětí svobody...`)
+            .setThumbnail('https://upload.wikimedia.org/wikipedia/commons/thumb/8/8d/Paper-notes.svg/512px-Paper-notes.svg.png') //Obrázek vlevo
+            .addFields(
+                { name: '**OSOBA**', value: jmeno_promene[1], inline: false },
+                { name: '**VÝŠE POKUTY**', value: jmeno_promene[2], inline: false },
+                { name: '**DŮVOD**', value: jmeno_promene[3], inline: false },
+            )          
+            .setTimestamp()
+            //.setFooter(`Pokutu vystavil ${message.member.nickname} a zapsal jí do databáze PČR`, 'https://media.discordapp.net/attachments/729654155798380548/769906915022471208/pcr.png');           // URL obrázku dole
         message.channel.send(exampleEmbed)
         message.delete()
     }
@@ -133,7 +152,7 @@ client.on('message', message => {
             .setAuthor('Policie České republiky', 'https://media.discordapp.net/attachments/729654155798380548/769906915022471208/pcr.png')    //Horní část embedu - Kdo to vytvoři, obrázek vpravu, co se otevře po kliknutí
             .setDescription(`Policista **${message.member.nickname}** si potvrdil kartou příchod do práce a vzal si výbavu.`)                //Text embedu
             .setTimestamp()
-            .setFooter(`Informace se zapsala do soukromé databáze Policie České republiky`, 'https://media.discordapp.net/attachments/729654155798380548/769906915022471208/pcr.png'           // URL obrázku dole
+            .setFooter(`Informace se zapsala do soukromé databáze Policie České republiky`, 'https://media.discordapp.net/attachments/729654155798380548/769906915022471208/pcr.png')           // URL obrázku dole
         message.channel.send(exampleEmbed).then(message.member.roles.add('769911778350989342'))
         message.delete()
     }
